User Management, Role Management, Permission Management Change. Sistem Kinerja Pegawai Feature.

+++ resources/views/users/index.blade.php
@@ -20,9 +20,10 @@
 
                                 <thead>
                                     <tr>
-                                        <th>Name</th>
+                                        <th>NIP</th>
+                                        <th>Nama</th>
                                         <th>Email</th>
-                                        
+                                        <th>Jabatan</th>
                                         <th>User Roles</th>
                                         <th>Operations</th>
                                     </tr>
@@ -31,15 +32,15 @@
                                 <tbody>
                                     @foreach ($users as $user)
                                     <tr>
-
+                                        <td>{{ $user->nip }}</td>
                                         <td>{{ $user->name }}</td>
                                         <td>{{ $user->email }}</td>
-                                      
+                                        <td>{{ $user->jabatan }}</td>
                                         <td>{{  $user->roles()->pluck('name')->implode(' ') }}</td>{{-- Retrieve array of roles associated to a user and convert to string --}}
                                         <td>
                                         <a href="{{ route('users.edit', $user->id) }}" class="btn btn-info pull-left" style="margin-right: 3px;">Edit</a>
 
-                                        {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id] ]) !!}
+                                        {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'onsubmit' => 'return confirm("Yakin menghapus data ini ('. $user->name . ')?")']) !!}
                                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                         {!! Form::close() !!}
